create functions to find missing experiments and bind into metaexperiment
Create functions to (i) find unique problems/solvers in a list of experiments, (ii) find missing problem solver pairs in a list of experiments, and (iii) construct a complete metaexperiment given a partial set of experiments. __init__() methods of Experiment and MetaExperiment were modified to allow for multiple ways to create these objects.