Review of Assemble
I reformatted the .rst file because the variables were not being expressed in mathematical notation and were unclear to read. They should display appropriately now.

The import functions of the model were not formatted correctly which caused a bunch of trickle down issues, but that is now fixed as well.

A variable later referenced as "machines_q" was defined as "machine_q" which caused problems but that is fixed now

A number of syntax errors specifically in the init function prevented demo_model from being run. These were fixed.

arrival_times_rng and product_orders_rng are called but never defined with an rng list. This should be further investigated to determine if they need to be defined as such or if they are not rng lists. In either case, these variables lead to errors as is.

The “item_cap” model factor is a list of lists and its check function does not work properly. I think this factor can be replaced with just a list of ints, and its check function would work properly if this is the case

The check_proccess_time function does not work due to a similar problem as the check_item_cap function, but this factor can not be changed as easily so the check function itself will just need to be changed to accommodate a list of lists.

The problem class for this model still had some references to the Hotel model that this code was seemingly based on. The problem class should be further analyzed to ensure it is accurate to the assembly model and not the hotel model.