Added logic for mu=0 edgecase
Theoretically mu should never be passed 0, but it's happening in one of the tests (which causes a divide by 0 error later on). To avoid this, I'm replacing 0 with the min float (as close to 0 as possible without being 0). This should have a minimal effect on the end result while avoiding the DB0 error.