Split tests up into different python files
Allows for the use of integrated test runners (such as the test runner built into VS Code) and allows for single tests to be ran easily. Also, each method (run, post_replicate, post_normalize) is now its own test and can be executed without running the other methods. This lets us isolate errors to the method in which they occurred instead of the test ending after a step in any of the three methods.